requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
dnspython==2.7.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.0
python-dotenv==1.1.1
typing-inspection==0.4.1
typing_extensions==4.14.1
Werkzeug==3.1.3
